requirements.txt
mcp[cli]>=1.26.0
plaid-python>=38.4.0
python-dotenv>=1.0.0
flask>=3.0.0
